KTOR-1915 Move appendBlob to web source set for shared JS/WasmJS code
Unify the appendBlob API across JS and WasmJS by placing the public
functions in the web source set. Both platforms now use org.w3c.files.Blob
from kotlinx-browser. Platform-specific blob-to-channel conversion remains
in js/src and wasmJs/src as expect/actual implementations.
